corpus: the whole-corpus rebank — one deliberate clean regen at Go 1.23.1
Regenerates src/go-src-converted production sources with the current
converter (seeded regen, path+content-precise hand-owned gate: 28
marked, 0 clobbered), ending the stale-bank drift era. Carries every
landed emission class since the Phase-3 tree (element-address aliasing,
ValueSlot prologues, select operand hoists, typed-nil singletons,
managed reinterpret, TypeAccessibility sections, float-suffix, sstring,
Delta-renames, ...) plus the 14 refreshed .cs.auto review siblings and
sync.Locker's inline public (the one hand-owned interface the
accessibility section cannot cover). Gated before this commit: full
corpus compile 0 errors and a 43/43 serial operational sweep on exactly
this tree.

+++ src/go-src-converted/internal/buildcfg/cfg.cs
@@ -431,7 +431,7 @@ internal static slice<@string> gogoarchTags() {
         if (major == 9) {
             // ARM64 v9.x also includes support of v8.x+5 (i.e. v9.1 includes v8.(1+5) = v8.6).
             for (nint i = 0; i <= minor + 5 && i <= 9; i++) {
-                list = append(list, fmt.Sprintf("%s.v%d.%d"u8, GOARCH, 8, i));
+                list = append(list, fmt.Sprintf("%s.v%d.%d"u8, GOARCH, (nint)(8), i));
             }
         }
         return list;
